fingerprint the .git/HEAD file in all jobs
To track relationships between jobs, Jenkins use a md5 of files presents
in the job workspace which uniquely idenitfy the source files being used
and hence the version behind tested.
.git/HEAD will give us an unique id identifying the current checkout.

The path to the file must be relative to the job workspace and since we
use a shared one, we have to hack around using symbolic links.

Simple hack that still let us avoiding copying the git repository
everywhere around.

Change-Id: If577cc143ac0d272894839776302756272ffe554