repair MediaWiki jobs dependencies
To track dependencies between jobs, Jenkins uses a concept known as
fingerprinting.  Once a build has finished, Jenkins will make a md5sum
on a set of files you have defined. File age is used to find out who is
the parent apparently.

Since md5 is done at the end of the build, the git fetching job was not
considered the parent job which broke dependencies.  This issue got
fixed by installing the "create fingerprint" plugin which let us request
fingerprint as the very first task, making the git fetching considered
to be the parent job again.

Change-Id: I10ac97e6d53dd28c549a5234bb9656a8729e0080