checkstyle target to run PHP CodeSniffer analyses
The git-files-changed target should be listing the files changed by the
last checkout. That is the only one we want to check style for or we
will end up returning unrelated issues.

CodeSniffer use the PEAR format standard for now, which is nowhere near
what we actually do in MediaWiki codebase.

Change-Id: I1bb25a69dc202cc9e83df2bf288b33634cf43ab7