make sure to clean DB before installing one
All tests are somehow sharing the same Sqlite database for now. The
intent was to have tests to run as fast as possible.

Since we can build changes from different branches, we really want to
start over by using a fresh database. Also, this let us test the
installer as a side effect.

Change-Id: I26b34888f6defb6080e8e40c281a3225e60c6f5d