testswarm sqlite db is now writable by others
I though that using set-group-id on a directory would be enough to
ensure the SQLite databases would always belong to the www-data group
and thus writable by Apache.

I was wrong :-( Somehow the nice Java application (either ant or
jenkins) is not honoring the set-group-id bit and both the directory and
the sqlite file belong to jenkins group.

Being pragmatic, lets just allow anyone to write in the files.

Change-Id: If605b0c21c238bd7fb46e23a45ee89a567bc433f