`Gerrit-Set-Reviewers` job to add reviewers
We have that nice idea to let Jenkins automatically add reviewers to
Gerrit changes as soon as they are submitted.

The first use case would be to have the Wikimedia Foundation database
administrators to be warned whenever a change involve a database change.

See https://bugzilla.wikimedia.org/show_bug.cgi?id=36228 for more
details.

For now, this job is doing nothing but listening for change made to the
test repositories under `test/mediawiki/`.

Change-Id: I7670f51faa1b5f56d6e5d6251381114c5e4b9337