ant: make git macros to always use 'dir'
Under ant, you end up running all <exec> calls in the $builddir
directory which is almost never where you want to run the git command.
All macros are no requiring a dir attribute to be set, that is safeguard
to make sure we are running git in the correct working directory.

Also removed a useless echo in the git macrodef. Did not output the
arguments but just the git command.

Change-Id: I5aad2c4dabcd3e3da8d750cb035ae4c493c759ca