sane default for builddir / sourcedir
Most of the time, we are going to set the builddir to be the Jenkins job
workspace. So we might as well use that as a default instead of failing
the build early.
MediaWiki jobs are sharing the MediaWiki-GIT-Fetching workspace by
setting sourcedir. For other jobs, we just want to use the default
workspace which is usually where the sourcecode is fetched by the Git
plugin.

Change-Id: If77568a78104111786ca64a9e8b63244dca492dd