Ext-Wikibase now report back to Gerrit (non voting)
The job has been failing tests for age so it was run under silent mode
and never reporting back to Gerrit. The reason was to avoid blocking the
changes when Jenkins verified -1.  Per discussion with Daniel, we simply
alter the configuration so jenkins-bot does not vote at all but still
add the URL for people to look at it.

Change-Id: I6f582f51aa53e663f86f6bbcf5e5283a561ad21a