ant sourcedir now default to workspace
We had ant sourcedir default to the MediaWiki-GIT-Fetching job workspace
which was fine while we only had a few tests that were all sharing the
same source directory. Since we are adding more and more extensions that
uses a copy of MediaWiki for their tests, we might as well use the job
workspace as a default sourcedir value.

Change-Id: Ic431b9a0dace36a62040729ca6b1c739a6533da9