apache debug log permission need to be done by Job
I attempted to fix the Apache debug log permission but forgot the code
path was being run by Apache.  It does not have the permission to fix
the permission, that needs to be done by jenkins-slave user.

Will update the JJB qunit macro so it creates the file for us and apply
a world writable permission.

Change-Id: I005e783b54a24226aa6fd0158f67c6675f0afa13