multigit: handles multiple git repository
Whenever a job depends on multiple repository, the Jenkins git plugin
would not let us apply the Gerrit patch on the proper repository.

The multgit.sh is adapted from OpenStack devstack-gate script. The idea
is to pass as parameters the Gerrit project to clone (optionally with a
destination directory separated with ':').  Whenever a requested project
match the Change that triggered the build, we would apply the patch,
else we simply attempt to match the branch of the Change.

The use cases are jobs shared for:

- VisualEditor / Parsoid
- Parsoid / Parsoid/deploy

Change-Id: I01c6bb5dff5185670ec09ff300f278e7deea23bc