multigit: fetch branches from Gerrit, not Zuul
Would attempt to make repositories to use the same branch as the one
that triggered the patch. The script would fetch the branch from Zuul
which is most probably not the actual branch in the Gerrit repository.

Instead, make the origin repository either the local Gerrit replica or
the Gerrit distant repo.  The Zuul patches are still fetched from Zuul
repositories though.

bug: 59224
Change-Id: I6a602ab35accf4b407a4d6bbae9e26b47d840ddc