Dump corefiles on segfaults
When e.g. phpunit crashes, it's useful to have a corefile for postmortem
analysis, as sometimes crashes are hard to reproduce outside the
environment or when trying to catch heisenbugs.

Set a 64MB upper ulimit (it should suffice to get started with) and set
it globally on the included mw-set-env.sh script to be included by all
run scripts.

Bug: 62623
Change-Id: I7a79e8e02123251732f4c729e00928f9f55ae9a3