grunt: Make grunt basedir and Gruntfile.js parent dir the same
The new grunt-cli version no longer supports having gruntfile
in one dir and node_modules/grunt in a completely different path.

This actually makes sense and was an undocumented feature to begin
with that we didn't even intend to make use of. The files should
both be in tools/, just like package.json.

Keeping symlink just for paranoia.

Change-Id: I04d8e308daedb3ccb80d18b9a7ae96eb776bbbdf