JSHint: Upgrade from v2.4.4 to v2.5.0
Matching the version various projects are already
using in their package.json for npm-test.

This affects the projects not using npm-test and/or
the check pipeline.

This ensures that the test pipeline using the local
npm-test and package.json matches the check pipeline.

This is another reason why we don't want slave-scripts
and having to do global upgrades. Projects are much better
off handling this in their own repo so that we don't risk
breaking people's master builds, and so that we can run
them in a sandbox build environment.

Release notes:
* https://github.com/jshint/jshint/releases/tag/2.5.0

Change-Id: I4a4e2a4c90fb6395e1b929b7a70a14906bc22ff0