Sleep one second between update.php runs
update.php --schema insert an entry in updatelog such as:
 updatelist-1.24alpha-1403726256

When running update.php without --schema, it attempts to use the same
key if it started during the same second.  That causes a duplicate key:
Error: 19 column ul_key is not unique

Sleeping one second is lame but workaround the issue.

Change-Id: I4c09530fc81fccf2368b3ca7e795a29c34d421f3