mwconf: let us override magic extension loader
I am going to have to keep a job workspace with a bunch of extension
from a previous run.  So let us skip the magic autoloader by listing the
extensions we want to require in a file named $IP/extension_load.txt

The file will be generated by the job I am working on to test all
extensions together.

Change-Id: Iaa4f2aa2dc0efc4e69c9545dbc0df12fa54866b7