mw-api-siteinfo: raise on http error
Whenever the target API URl returns a 404 / 503, we would pass the
payload to json.loads() which gives a missleading error message.

Use response.raise_for_status() to have requests raise an exception
whenever the http request is deemed to have failed.

Tested by Antoine using requests 0.8.2 which is the version in use on
Precise instances.

Bug: 70695
Change-Id: Ic2a5eca1c0f2f449dfce8b8786417a56b58acbbb