Point Doxygen input filter to Zend PHP
On Trusty instances, 'php' refers to HHVM which has a rather large byte
code cache overhead.  That slow down the invocation of mwdoc-filter.php
from 40ms (Zend) to 200+ms (PHP).

Set $wgPhpCli to the Zend version of PHP ('/usr/bin/php5') so invocation
of wfShellWikiCmd() ends up using it when crafting the doxygen
INPUT_FILTER.

With HHVM the doc generation takes up to 20 minutes. With this patch we
are down to 6 minutes (without graph generation) which is slightly
better.

Bug: 73311
Change-Id: I48b62c9213faab2f114c41922484def8d90f9426