De-duplicate LOG_DIR logic
The compression trap has been removed. I was going to restore this in a postbuild
step in the  Jenkins jobs, but I don't think it's worth the overhead.
We only keep the build artefacts for a short amount of time anyway.

This centralises:
* definition of LOG_DIR.
* creation of LOG_DIR (if not already created).
* chmod permissinos for LOG_DIR.

It does not centralise emptying of the log directory since there
are some Jobs that ensure mw-set-env multiple times and this would
result in loss of logs.

Change-Id: I66fbce3fa0a8cd91da9c1ceb66ef224016cb247f