Copy LocalSettings.php to "/log" in teardown instead of setup
This way other modifications after mw-apply-settings, such as
mw-apply-wb-settings and macro/qunit, are also reflected in the
file collected by Jenkins build artefacts.

This would've helped figure out where the syntax error was in
the Wikibase build last month. The file uploaded to Jenkins was
from before the error was appended.

Bug: T90613
Change-Id: I52f81eae4d795ca03ccb0882fd012be1841ad061