Fix headless setup for MW-Selenium
There are a few bugs we have to workaround for now.

The first is a bug in the headless gem which prevents us from reusing
the existing Xvfb display so we'll start up Xvfb on an isolated display
for each build. This may actually be a preferable long-term setup in
case we need to take screen caps, etc.

https://github.com/leonid-shevtsov/headless/pull/60

Second is a bug in xorg on Ubuntu that prevents us from using a TMPDIR
on a second filesystem, so we'll force use of /tmp for now.

https://bugs.launchpad.net/ubuntu/+source/xorg-server/+bug/972324

Change-Id: Iaf3287cfb3fe897de6fb4c4b79784a5e7ef900a8