Rebuild tools packages for Ubuntu Trusty
Follows-up f5820781, e94262d6, 29d7023a.

All versions kept as-is, except:
* Bump phantomjs from 1.9.7-3 to 1.9.8.
* Remove grunt-contrib-csslint.

The installer of phantomjs 1.9.7-3 no longer works due to an
oudated CDN url. https://github.com/Medium/phantomjs/issues/310
Could not recompile without upgrading first. Version 1.9.8 is still
old, but that is intentional as this tools bundle is for
compatibility purposes only. (For check jobs.) This version has
been tested with MediaWiki, OOjs, and QUnit and known to be feature
compatible.

The grunt-contrib-csslint is no longer used in check jobs.
The new npm jobs have a local copy instead.

Other packages should not be upgraded as it would break existing
repositories' master build.

Instead, repositories should update their job config to not use
these scripts in the first place. Until then, this is updated for
maintenance and/or security reasons only.

Change-Id: I5881ff7c14dd757d0be196cdbcce423dd64dd028