Remove obsolete tools/fetch-mw-ext
The last usage was a JJB macro used by legacy browser tests jobs. We
came up with a better system and https://gerrit.wikimedia.org/r/234519
did the cleanup.

The script is no more used anywhere and can be removed. That removes a
dependency we had toward git.wikimedia.org.

Bug: T74001
Change-Id: Id05e2475030c46e7f7517045b6bf92c6c1543932