Add commit-message-validator.py tool
Add a python script that can be used to validate the format of a commit
message. When executed, the script will get the commit message of the
HEAD commit in the current working directory and validate it. Script
will exit with a non-zero status if validation fails. Validation errors
are written to stdout.

Current validation rules:
* First line <=80 characters
* Second line blank
* No line >100 characters
* "Bug:" is capitalized
* "Bug:" is followed by a space
* Exactly one task id on each Bug: line
* No "Task:", "Fixes:", "Closes:" lines
* For any "Bug: " line, next line is not blank
* For any "Bug: " line, prior line is another Bug: line or empty
* Exactly one "Change-Id: " line per commit
* For "Change-Id: " line, prior line is empty or "Bug: "
* No blank lines between any "Bug: " lines and "Change-Id: "
* Only "(cherry picked from commit" can follow "Change-Id: "
* Message has at least 3 lines (subject, blank, Change-Id)

Bug: T109119
Change-Id: Ib07957562a0f54abc669f91db9b8bddb29d369c2