[IMPROV] Test returned content first
Instead of just checking the exit code, it's more helpful to actually see the
output difference. For example when it excepts success it would only tell that
it got a failure but not what failed. If it expected a failure and would get
success it would show an empty string which implies that it erroneously
succeeded.

Change-Id: If384513b9f952b91f44dfd0b5a3144abb46f3be7