[IMPROV] Don't use shell argument
The `shell` argument in `subprocess.check_output` should be used very
carefully. But this is not necessary if we already split the arguments. This
now also uses a custom log format which will only print the commit message not
indented.

This also fixes various issues that it didn't detect lines which start with
"Change-Id" or "Bug" because it didn't consider the indentation.

Change-Id: I7724c6c75ed100ef11082e95f4f48ef3f5d91d6a