mw-apply-settings: set $wgTmpDirectory explicitly
When MediaWiki is hit via Apache, the env variable TMPDIR is not set and
$wgTmpDirectory and others ends up to fall back to /tmp.  The
localisation cache would eventually use that and since it is not garbage
collected between build that might cause side effects on subsequent
builds.

MW_TMPDIR is set via mw-set-env / global-set-env. Eval the value to set
explicitly set $wgTmpDirectory.

Bug: T120356
Change-Id: I6b852c5d8ab29f64ec3718dd9a659af7eb5c6806