Bug fix: Node services: fix up the module path in the config
service-runner, the library used for running WMF node services, needs
the path to the service's entry-point module. However, due to a quirk in
the library itself, when relative paths are provided, it tries to load
them by using the parent directory of node_modules as the root prefix.
This works when run inside source repositories, but for deploy
repositories the path differs. This commit works around that by fixing
up the config file and providing the full path to the service's module
file.

Also in this commit: print out the NPM logs when installing the
development node modules, as they are known to contain binary
dependencies which might fail to install sometimes.

Note that the script now uses the NPM_SET_PATH environment variable to
detect the source repo's location and does no config massaging if the
config.yaml file is not present in it, and is hence compatible with
non-service-runner services as well (i.e. Parsoid).

Change-Id: I0b8f5e490fe8db67333adc763d00edde418d0fca