npm-install-dev.py: Use config.dev.yaml instead of config.yaml
We have counted on the fact that all of the repos have a config.yaml
file which is a symlink to config.dev.yaml. However, this was not a good
approach because the tests are always run against config.dev.yaml,
causing them to fail for repos where config.yaml is either missing or a
proper file.

Bug: T124447
Change-Id: I2db6327657021c80a8826b472e7dabb2038651ed