Add run-commit-message-validator script
This creates a virtualenv, installs the script, and executes it.

Typically we'd encourage people use to use `tox` for Python lint checks,
but since this is specific to git commit messages, we'd probably want
people to use a git commit hook rather than manually running tox.

Also, per-repository versioning is not an issue here as even though the
commit message guidelines is a living document, we'd want people to
automatically upgrade. Probably. We'll see.

Change-Id: I24e6bb5631ced817323fb9bbba684fabda0fb58c