Get rid of manual PHP_BIN handling
The original implementation of PHP_BIN involved replacing each
invocation of `php` with `$PHP_BIN`, but we soon switched to using the
altneratives system, with a wrapper that picked the right PHP version
from the environment variable.

This cleans up after that initial implementation by replacing the manual
$PHP_BIN invocations with `php` so everything goes through one,
consistent, entry-point.

Change-Id: Ifa800aa640049460cff186e9cf570fc1855e7f45