Option to merge mw composer into vendor
PHP is apparently  not being to lookup a static method through the
namespaces.  Eg the following does not work:

    use Hamcrest\Matchers;
    both( .. );

For static methods, PHP requires to use at least the class, eg with the
full canonical path it works fine:

    \Hamcrest\Matchers::both( ... );

hamcrest/hamcrest-php and wmde/hamcrest-html-matchers have a syntastic
sugar to expose some of their static methods as global functions. They
can be loaded either explicitly with:

    Hamcrest\Util::registerGlobalFunctions()

Or using a composer files autoloader which we have done in
mediawiki/core:

    "autoload-dev": {
        "files": [
            "vendor/hamcrest/hamcrest-php/hamcrest/Hamcrest.php",
            "vendor/wmde/hamcrest-html-matchers/src/functions.php"
        ]
    },

In the first example, both() would be resolved by PHP as the global
function it get invoked and ends up invoking the static method. The code
works.

When using mediawiki/vendor, mw-fetch-composer-dev installs the dev
dependencies (via composer-dev-args.js) but the autoload-dev section is
NOT merged in. The global functions are hence not registered and the
build fails.

Add option MW_COMPOSER_MERGE_MW_IN_VENDOR which would add in the
vendor/composer.json an extra section for the composer-merge-plugin
instructing it to merge in mediawiki/core.  That will add the
require-dev sections (done previously) and the autoload-dev section.

That is done before 'composer dump-autoload' which would then generate
a vendor/composer/autoload_file.php containing the files from
mediawiki/core autoloader-dev section.

Also copy (existing) resulting files to log dir so they get attached to
the build.

Bug: T158674
Change-Id: I5fd7411560aed93a5578a8c0cf5cdcc73384d9b5